+++ app/src/main/java/com/armoz/roadtoalcatraz/base/domain/model/PlayModel.java
@@ -4,14 +4,14 @@ package com.armoz.roadtoalcatraz.base.domain.model;
  * Created by ruben.arana on 24/12/15.
  */
 public class PlayModel {
-    private long playerId;
+    private int playerId;
     private int action;
 
-    public long getPlayerId() {
+    public int getPlayerId() {
         return playerId;
     }
 
-    public void setPlayerId(long playerId) {
+    public void setPlayerId(int playerId) {
         this.playerId = playerId;
     }
